requirements.txt
python-telegram-bot==13.15
yfinance==0.1.64
SQLAlchemy==1.4.25
requests==2.26.0
pymysql
streamlit